Applied fix to handle_read which ensures that data is read at a reasonable rate.  Taken from an internal version of the module in use at CCP Games, with permission.
On Windows, we were seeing local transfers at the rate of 16 k/s and remote transfers at the rate of 700 k/s.  The difference appeared to be the size of the chunks in which the sent data was delivered.  Reading more than one chunk for each handle_read call accounts for the difference in chunk size, and allows both local and remote transfers to happen at the same maximised rate.